+++ cpp/include/cuvs/distance/distance.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024, NVIDIA CORPORATION.
+ * Copyright (c) 2021-2023, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,12 +16,53 @@
 
 #pragma once
 
+#include "distance.h"
+
 #include <cstdint>
-#include <cuvs/distance/distance_types.hpp>
 #include <raft/core/device_mdspan.hpp>
 #include <raft/core/resources.hpp>
 
 namespace cuvs::distance {
+
+using DistanceType = cuvsDistanceType;
+
+/**
+ * Whether minimal distance corresponds to similar elements (using the given metric).
+ */
+inline bool is_min_close(DistanceType metric)
+{
+  bool select_min;
+  switch (metric) {
+    case DistanceType::InnerProduct:
+      // Similarity metrics have the opposite meaning, i.e. nearest neighbors are those with larger
+      // similarity (See the same logic at cpp/include/raft/sparse/spatial/detail/knn.cuh:362
+      // {perform_k_selection})
+      select_min = false;
+      break;
+    default: select_min = true;
+  }
+  return select_min;
+}
+
+namespace kernels {
+enum KernelType { LINEAR, POLYNOMIAL, RBF, TANH };
+
+/**
+ * Parameters for kernel matrices.
+ * The following kernels are implemented:
+ * - LINEAR \f[ K(x_1,x_2) = <x_1,x_2>, \f] where \f$< , >\f$ is the dot product
+ * - POLYNOMIAL \f[ K(x_1, x_2) = (\gamma <x_1,x_2> + \mathrm{coef0})^\mathrm{degree} \f]
+ * - RBF \f[ K(x_1, x_2) = \exp(- \gamma |x_1-x_2|^2) \f]
+ * - TANH \f[ K(x_1, x_2) = \tanh(\gamma <x_1,x_2> + \mathrm{coef0}) \f]
+ */
+struct KernelParams {
+  // Kernel function parameters
+  KernelType kernel;  //!< Type of the kernel function
+  int degree;         //!< Degree of polynomial kernel (ignored by others)
+  double gamma;       //!< multiplier in the
+  double coef0;       //!< additive constant in poly and tanh kernels
+};
+}  // end namespace kernels
 
 /**
  * @defgroup pairwise_distance Pairwise Distances API
@@ -37,7 +78,7 @@ namespace cuvs::distance {
  * @code{.cpp}
  * #include <raft/core/resources.hpp>
  * #include <raft/core/device_mdarray.hpp>
- * #include <cuvs/distance/pairwise_distance.hpp>
+ * #include <cuvs/distance/distance.hpp>
  *
  * raft::resources handle;
  * int n_samples = 5000;
@@ -81,7 +122,7 @@ void pairwise_distance(
  * @code{.cpp}
  * #include <raft/core/resources.hpp>
  * #include <raft/core/device_mdarray.hpp>
- * #include <cuvs/distance/pairwise_distance.hpp>
+ * #include <cuvs/distance/distance.hpp>
  *
  * raft::resources handle;
  * int n_samples = 5000;
@@ -125,7 +166,7 @@ void pairwise_distance(
  * @code{.cpp}
  * #include <raft/core/resources.hpp>
  * #include <raft/core/device_mdarray.hpp>
- * #include <cuvs/distance/pairwise_distance.hpp>
+ * #include <cuvs/distance/distance.hpp>
  *
  * raft::resources handle;
  * int n_samples = 5000;
@@ -168,7 +209,7 @@ void pairwise_distance(
  * @code{.cpp}
  * #include <raft/core/resources.hpp>
  * #include <raft/core/device_mdarray.hpp>
- * #include <cuvs/distance/pairwise_distance.hpp>
+ * #include <cuvs/distance/distance.hpp>
  *
  * raft::resources handle;
  * int n_samples = 5000;
@@ -205,4 +246,4 @@ void pairwise_distance(
 
 /** @} */  // end group pairwise_distance_runtime
 
-}  // namespace cuvs::distance
+};  // namespace cuvs::distance
