--- cpp/include/cuvs/neighbors/sample_filter.hpp
@@ -1,171 +1,0 @@
-/*
- * Copyright (c) 2023-2024, NVIDIA CORPORATION.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-#pragma once
-
-#include <cuvs/core/bitset.hpp>
-#include <raft/core/detail/macros.hpp>
-
-#include <cstddef>
-#include <cstdint>
-
-namespace cuvs::neighbors::filtering {
-
-/* A filter that filters nothing. This is the default behavior. */
-struct none_ivf_sample_filter {
-  inline _RAFT_HOST_DEVICE bool operator()(
-    // query index
-    const uint32_t query_ix,
-    // the current inverted list index
-    const uint32_t cluster_ix,
-    // the index of the current sample inside the current inverted list
-    const uint32_t sample_ix) const;
-};
-
-/* A filter that filters nothing. This is the default behavior. */
-struct none_cagra_sample_filter {
-  inline _RAFT_HOST_DEVICE bool operator()(
-    // query index
-    const uint32_t query_ix,
-    // the index of the current sample
-    const uint32_t sample_ix) const;
-};
-
-/**
- * @brief Filter used to convert the cluster index and sample index
- * of an IVF search into a sample index. This can be used as an
- * intermediate filter.
- *
- * @tparam index_t Indexing type
- * @tparam filter_t
- */
-template <typename index_t, typename filter_t>
-struct ivf_to_sample_filter {
-  const index_t* const* inds_ptrs_;
-  const filter_t next_filter_;
-
-  ivf_to_sample_filter(const index_t* const* inds_ptrs, const filter_t next_filter);
-
-  /** If the original filter takes three arguments, then don't modify the arguments.
-   * If the original filter takes two arguments, then we are using `inds_ptr_` to obtain the sample
-   * index.
-   */
-  inline _RAFT_HOST_DEVICE bool operator()(
-    // query index
-    const uint32_t query_ix,
-    // the current inverted list index
-    const uint32_t cluster_ix,
-    // the index of the current sample inside the current inverted list
-    const uint32_t sample_ix) const;
-};
-
-/**
- * @brief Filter an index with a bitset
- *
- * @tparam index_t Indexing type
- */
-template <typename bitset_t, typename index_t>
-struct bitset_filter {
-  // View of the bitset to use as a filter
-  const cuvs::core::bitset_view<bitset_t, index_t> bitset_view_;
-
-  bitset_filter(const cuvs::core::bitset_view<bitset_t, index_t> bitset_for_filtering);
-  inline _RAFT_HOST_DEVICE bool operator()(
-    // query index
-    const uint32_t query_ix,
-    // the index of the current sample
-    const uint32_t sample_ix) const;
-};
-
-/**
- * If the filtering depends on the index of a sample, then the following
- * filter template can be used:
- *
- * template <typename IdxT>
- * struct index_ivf_sample_filter {
- *   using index_type = IdxT;
- *
- *   const index_type* const* inds_ptr = nullptr;
- *
- *   index_ivf_sample_filter() {}
- *   index_ivf_sample_filter(const index_type* const* _inds_ptr)
- *       : inds_ptr{_inds_ptr} {}
- *   index_ivf_sample_filter(const index_ivf_sample_filter&) = default;
- *   index_ivf_sample_filter(index_ivf_sample_filter&&) = default;
- *   index_ivf_sample_filter& operator=(const index_ivf_sample_filter&) = default;
- *   index_ivf_sample_filter& operator=(index_ivf_sample_filter&&) = default;
- *
- *   inline _RAFT_HOST_DEVICE bool operator()(
- *       const uint32_t query_ix,
- *       const uint32_t cluster_ix,
- *       const uint32_t sample_ix) const {
- *     index_type database_idx = inds_ptr[cluster_ix][sample_ix];
- *
- *     // return true or false, depending on the database_idx
- *     return true;
- *   }
- * };
- *
- * Initialize it as:
- *   using filter_type = index_ivf_sample_filter<idx_t>;
- *   filter_type filter(cuvs_ivfpq_index.inds_ptrs().data_handle());
- *
- * Use it as:
- *   cuvs::neighbors::ivf_pq::search_with_filtering<data_t, idx_t, filter_type>(
- *     ...regular parameters here...,
- *     filter
- *   );
- *
- * Another example would be the following filter that greenlights samples according
- * to a contiguous bit mask vector.
- *
- * template <typename IdxT>
- * struct bitmask_ivf_sample_filter {
- *   using index_type = IdxT;
- *
- *   const index_type* const* inds_ptr = nullptr;
- *   const uint64_t* const bit_mask_ptr = nullptr;
- *   const int64_t bit_mask_stride_64 = 0;
- *
- *   bitmask_ivf_sample_filter() {}
- *   bitmask_ivf_sample_filter(
- *       const index_type* const* _inds_ptr,
- *       const uint64_t* const _bit_mask_ptr,
- *       const int64_t _bit_mask_stride_64)
- *       : inds_ptr{_inds_ptr},
- *         bit_mask_ptr{_bit_mask_ptr},
- *         bit_mask_stride_64{_bit_mask_stride_64} {}
- *   bitmask_ivf_sample_filter(const bitmask_ivf_sample_filter&) = default;
- *   bitmask_ivf_sample_filter(bitmask_ivf_sample_filter&&) = default;
- *   bitmask_ivf_sample_filter& operator=(const bitmask_ivf_sample_filter&) = default;
- *   bitmask_ivf_sample_filter& operator=(bitmask_ivf_sample_filter&&) = default;
- *
- *   inline _RAFT_HOST_DEVICE bool operator()(
- *       const uint32_t query_ix,
- *       const uint32_t cluster_ix,
- *       const uint32_t sample_ix) const {
- *     const index_type database_idx = inds_ptr[cluster_ix][sample_ix];
- *     const uint64_t bit_mask_element =
- *         bit_mask_ptr[query_ix * bit_mask_stride_64 + database_idx / 64];
- *     const uint64_t masked_bool =
- *         bit_mask_element & (1ULL << (uint64_t)(database_idx % 64));
- *     const bool is_bit_set = (masked_bool != 0);
- *
- *     return is_bit_set;
- *   }
- * };
- */
-}  // namespace cuvs::neighbors::filtering
