+++ cpp/src/neighbors/brute_force.cu
@@ -26,7 +26,7 @@ index<T>::index(raft::resources const& res,
                 std::optional<raft::device_vector<T, int64_t>>&& norms,
                 cuvs::distance::DistanceType metric,
                 T metric_arg)
-  : ann::index(),
+  : cuvs::neighbors::index(),
     metric_(metric),
     dataset_(raft::make_device_matrix<T, int64_t>(res, 0, 0)),
     norms_(std::move(norms)),
@@ -43,7 +43,7 @@ index<T>::index(raft::resources const& res,
                 std::optional<raft::device_vector<T, int64_t>>&& norms,
                 cuvs::distance::DistanceType metric,
                 T metric_arg)
-  : ann::index(),
+  : cuvs::neighbors::index(),
     metric_(metric),
     dataset_(raft::make_device_matrix<T, int64_t>(res, 0, 0)),
     norms_(std::move(norms)),
@@ -59,7 +59,7 @@ index<T>::index(raft::resources const& res,
                 std::optional<raft::device_vector_view<const T, int64_t>> norms_view,
                 cuvs::distance::DistanceType metric,
                 T metric_arg)
-  : ann::index(),
+  : cuvs::neighbors::index(),
     metric_(metric),
     dataset_(raft::make_device_matrix<T, int64_t>(res, 0, 0)),
     dataset_view_(dataset_view),
