+++ cpp/src/neighbors/cagra.cuh
@@ -28,7 +28,7 @@
 #include <cuvs/distance/distance_types.hpp>
 #include <cuvs/neighbors/cagra.hpp>
 
-#include <cuvs/neighbors/dataset.hpp>
+#include <cuvs/neighbors/common.hpp>
 
 #include <rmm/cuda_stream_view.hpp>
