+++ cpp/src/neighbors/detail/cagra/cagra_search.cuh
@@ -34,7 +34,7 @@
 // TODO: Fix these when ivf methods are moved over
 #include "../../ivf_common.cuh"
 #include "../../ivf_pq/ivf_pq_search.cuh"
-#include <cuvs/neighbors/sample_filter.hpp>
+#include <cuvs/neighbors/common.hpp>
 
 // TODO: This shouldn't be calling spatial/knn apis
 #include "../ann_utils.cuh"
