+++ cpp/src/neighbors/detail/cagra/search_multi_cta_kernel-inl.cuh
@@ -31,7 +31,7 @@
 
 #include <cuvs/distance/distance_types.hpp>
 
-#include <cuvs/neighbors/sample_filter.hpp>
+#include <cuvs/neighbors/common.hpp>
 
 // TODO: This shouldn't be invoking anything from spatial/knn
 #include "../ann_utils.cuh"
