+++ cpp/src/neighbors/detail/cagra/search_multi_kernel.cuh
@@ -32,7 +32,7 @@
 
 #include <raft/matrix/select_k.cuh>
 
-#include <cuvs/neighbors/sample_filter.hpp>
+#include <cuvs/neighbors/common.hpp>
 
 // TODO: This shouldn't be invoking anything from spatial/knn
 #include "../ann_utils.cuh"
