+++ cpp/src/neighbors/ivf_flat/ivf_flat_search.cuh
@@ -18,10 +18,10 @@
 
 #include "../../core/nvtx.hpp"
 #include "../detail/ann_utils.cuh"
-#include "../ivf_common.cuh"                 // cuvs::neighbors::detail::ivf
-#include "ivf_flat_interleaved_scan.cuh"     // interleaved_scan
-#include <cuvs/neighbors/ivf_flat.hpp>       // raft::neighbors::ivf_flat::index
-#include <cuvs/neighbors/sample_filter.hpp>  // none_ivf_sample_filter
+#include "../ivf_common.cuh"              // cuvs::neighbors::detail::ivf
+#include "ivf_flat_interleaved_scan.cuh"  // interleaved_scan
+#include <cuvs/neighbors/common.hpp>      // none_ivf_sample_filter
+#include <cuvs/neighbors/ivf_flat.hpp>    // raft::neighbors::ivf_flat::index
 
 #include "../detail/ann_utils.cuh"           // utils::mapping
 #include <cuvs/distance/distance_types.hpp>  // is_min_close, DistanceType
