+++ cpp/src/neighbors/ivf_list.cuh
@@ -16,7 +16,7 @@
 
 #pragma once
 
-#include <cuvs/neighbors/ivf_list.hpp>
+#include <cuvs/neighbors/common.hpp>
 #include <cuvs/neighbors/ivf_pq.hpp>
 
 #include <raft/core/device_mdarray.hpp>
