+++ cpp/src/neighbors/ivf_pq/ivf_pq_build.cuh
@@ -22,7 +22,7 @@
 #include "ivf_pq_codepacking.cuh"
 #include "ivf_pq_list.cuh"
 #include <cuvs/distance/distance_types.hpp>
-#include <cuvs/neighbors/ivf_list.hpp>
+#include <cuvs/neighbors/common.hpp>
 #include <cuvs/neighbors/ivf_pq.hpp>
 
 #include "../detail/ann_utils.cuh"  // utils::mapping
