+++ cpp/test/distance/distance_base.cuh
@@ -18,11 +18,11 @@
 
 #include "../test_utils.cuh"
 
+#include <cuvs/distance/distance.hpp>
 #include <cuvs/distance/distance_types.hpp>  // cuvs::distance::DistanceType
-#include <cuvs/distance/pairwise_distance.hpp>
-#include <raft/common/nvtx.hpp>         // raft::common::nvtx::range
-#include <raft/core/device_mdspan.hpp>  //raft::make_device_matrix_view
-#include <raft/core/operators.hpp>      // raft::sqrt
+#include <raft/common/nvtx.hpp>              // raft::common::nvtx::range
+#include <raft/core/device_mdspan.hpp>       //raft::make_device_matrix_view
+#include <raft/core/operators.hpp>           // raft::sqrt
 #include <raft/core/resource/cuda_stream.hpp>
 #include <raft/core/resources.hpp>  // raft::resources
 #include <raft/random/rng.cuh>
